[Issue #215] Move index files into subdirectories
