--- src/filters.ts
@@ -1,3 +1,0 @@
-export { getFilter } from './filters/get-filter';
-export { postFilter } from './filters/post-filter';
-export { patchFilter } from './filters/patch-filter';
