--- src/guards.ts
@@ -1,5 +1,0 @@
-export { onlyActive } from './guards/only-active';
-export { onlyAdmin } from './guards/only-admin';
-export { onlyMember } from './guards/only-member';
-export { onlySelf } from './guards/only-self';
-export { onlyUser } from './guards/only-user';
