--- src/middleware.ts
@@ -1,14 +1,0 @@
-/**
- * # Middleware
- *
- * Middleware are additional prepackaged functions to chain into your requests
- */
-
-/**
- * Middleware
- */
-// TODO: Remove in v4 vvv
-export { getQuery } from './query-tools/get-query';
-export { loadEntity } from './middleware/load-entity';
-export { loadUser } from './middleware/load-user';
-export { basicCors } from './middleware/basic-cors';
