+++ src/middleware/index.ts
@@ -1,0 +1,14 @@
+/**
+ * # Middleware
+ *
+ * Middleware are additional prepackaged functions to chain into your requests
+ */
+
+/**
+ * Middleware
+ */
+// TODO: Remove in v4 vvv
+export { getQuery } from '../query-tools/get-query';
+export { loadEntity } from './load-entity';
+export { loadUser } from './load-user';
+export { basicCors } from './basic-cors';
