+++ src/models/index.ts
@@ -41,9 +41,9 @@
 /**
  * Models
  */
-export { BaseModel, BaseModelInterface } from './models/base-model';
-export { BaseUser, BaseUserInterface } from './models/base-user';
-export { ExampleUser } from './models/example-user';
+export { BaseModel, BaseModelInterface } from './base-model';
+export { BaseUser, BaseUserInterface } from './base-user';
+export { ExampleUser } from './example-user';
 
 export function getISOTime() {
 	return new Date(Date.now()).toISOString();
