--- src/scripts.ts
@@ -1,10 +1,0 @@
-/**
- * # Scripts
- *
- * Script helper functions
- */
-export { bootstrap } from './scripts/bootstrap';
-export { bundle } from './scripts/bundle';
-export { bundlePackageJson } from './scripts/bundle-package-json';
-export { configure } from './scripts/configure';
-export { getBuildDirectory } from './scripts/get-build-directory';
