+++ src/scripts/index.ts
@@ -1,0 +1,10 @@
+/**
+ * # Scripts
+ *
+ * Script helper functions
+ */
+export { bootstrap } from './bootstrap';
+export { bundle } from './bundle';
+export { bundlePackageJson } from './bundle-package-json';
+export { configure } from './configure';
+export { getBuildDirectory } from './get-build-directory';
