--- src/utils.ts
@@ -1,16 +1,0 @@
-export { addResource } from './utils/add-resource';
-export { bindResponders } from './utils/bind-responders';
-// TODO: Remove in v4 vvv
-export { createSearchQuery } from './query-tools/create-search-query';
-export { createTimestamp } from './utils/create-timestamp';
-export { deleteUndefinedMembers } from './utils/delete-undefined-members';
-export { forkServer } from './utils/fork-server';
-export { getIdentifierValue } from './utils/get-identifier-value';
-export { isAdmin } from './utils/is-admin';
-export { isDefined } from './utils/is-defined';
-export { isJson } from './utils/is-json';
-export { isKeyInModel } from './utils/is-key-in-model';
-export { isSelf } from './utils/is-self';
-export { listen } from './utils/listen';
-export { runHook } from './utils/run-hook';
-export { upgradeUserRole } from './utils/upgrade-user-role';
